# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hoddmímir API",
    "version": "1.0.0",
    "description": "Authenticated, role-protected API for read projections and verified Hoddmímir configuration mutations."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "sessionCookie": [] }],
  "paths": {
...
    "/api/v1/inventory/overview": {
      "get": {
        "operationId": "getInventoryOverview",
        "responses": {

          "503": { "$ref": "#/components/responses/MaintenanceUnavailable" },
          "200": {
            "description": "Current inventory counts and UTC freshness.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InventoryOverview" }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/resources": {
      "get": {
        "operationId": "listInventoryResources",
        "parameters": [
          { "$ref": "#/components/parameters/Kind" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/ConnectionId" },
          { "$ref": "#/components/parameters/ParentId" },
          { "$ref": "#/components/parameters/InventoryState" },
          { "$ref": "#/components/parameters/GuestType" }
        ],
        "responses": {
          "503": { "$ref": "#/components/responses/MaintenanceUnavailable" },
          "200": {
            "description": "A bounded keyset page of inventory resources.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryResourcePage"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" }
        }
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthLoginRequest" }
            }
          }
        },
        "responses": {
          "503": { "$ref": "#/components/responses/MaintenanceUnavailable" },
          "200": {
            "description": "Authenticated local session.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthSession" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/AuthenticationRequired" }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getAuthSession",
        "responses": {
          "503": { "$ref": "#/components/responses/MaintenanceUnavailable" },
          "200": {
            "description": "Current principal and derived CSRF token; never the session token.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthSession" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/AuthenticationRequired" },
          "403": { "$ref": "#/components/responses/PermissionDenied" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "503": { "$ref": "#/components/responses/MaintenanceUnavailable" },
          "204": { "description": "Session revoked." },
          "401": { "$ref": "#/components/responses/AuthenticationRequired" },
          "403": { "$ref": "#/components/responses/PermissionDenied" }
        }
      }
    },
...
    "/api/v1/collector/status": {
      "get": {
        "operationId": "getCollectorStatus",
        "responses": {
          "503": { "$ref": "#/components/responses/MaintenanceUnavailable" },
          "200": {
            "description": "Collector schedule and latest heartbeat without lease identity or credentials.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CollectorStatus" }
              }
            }
          }
        }
      }
    },
    "/api/v1/collector/runs": {
      "get": {
        "operationId": "listCollectorRuns",
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Cursor" }
        ],
        "responses": {
          "503": { "$ref": "#/components/responses/MaintenanceUnavailable" },
          "200": {
            "description": "Latest bounded keyset page of inventory-sync runs.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CollectorRunPage" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" }
        }
      }
    },
    "/api/v1/collector/scopes": {
      "get": {
        "operationId": "listCollectorScopes",
        "parameters": [
          { "$ref": "#/components/parameters/RunId" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Cursor" }
        ],
        "responses": {
          "503": { "$ref": "#/components/responses/MaintenanceUnavailable" },
          "200": {
            "description": "Bounded keyset page of parent and PBS-content scope results for one inventory-sync run.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CollectorScopePage" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hoddmimir_session"
      }
    },
    "parameters": {
      "Kind": {
        "name": "kind",
        "in": "query",
        "required": true,
        "schema": { "$ref": "#/components/schemas/InventoryResourceKind" }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/PageCursor" }
      },
      "ConnectionId": {
        "name": "connectionId",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
      },
...
      "ParentId": {
        "name": "parentId",
        "in": "query",
        "description": "Cluster, PBS server, datastore/parent namespace, namespace, or backup-group ID as required by the selected kind.",
        "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
      },
      "InventoryState": {
        "name": "inventoryState",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/InventoryState" }
      },
      "GuestType": {
        "name": "guestType",
        "in": "query",
        "schema": { "type": "string", "enum": ["qemu", "lxc"] }
      },
      "RunId": {
        "name": "runId",
        "in": "query",
        "required": true,
        "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
      },
...
    },
    "responses": {
      "MaintenanceUnavailable": {
        "description": "Persistent deployment maintenance temporarily blocks application requests; retry after the Retry-After interval.",
        "headers": { "Retry-After": { "schema": { "type": "string" } } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MaintenanceError" } } }
      },
      "AuthenticationRequired": {
        "description": "A valid local web session is required.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AuthError" }
          }
        }
      },
      "PermissionDenied": {
        "description": "Permission or CSRF validation failed.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AuthError" }
          }
        }
      },
      "InvalidQuery": {
        "description": "Unknown, malformed, incompatible, or out-of-bounds query.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
...
    },
    "schemas": {
...
      "MaintenanceError": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": {
          "type": "object", "additionalProperties": false, "required": ["code", "message"],
          "properties": {
            "code": { "const": "maintenance" },
            "message": { "const": "Hoddmímir wird gewartet." }
          }
        } }
      },
...
      "AuthLoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 3, "maxLength": 64 },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 1,
            "maxLength": 4096
          }
        }
      },
      "AuthPrincipal": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "username", "permissions"],
        "properties": {
          "id": { "$ref": "#/components/schemas/CanonicalUuid" },
          "username": { "type": "string", "minLength": 3, "maxLength": 64 },
          "permissions": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "inventory.read",
                "backup_configuration.manage",
                "backup_operations.manage",
                "audit.read",
                "security.manage"
              ]
            }
          }
        }
      },
      "AuthSession": {
        "type": "object",
        "additionalProperties": false,
        "required": ["user", "csrfToken", "idleExpiresAt", "absoluteExpiresAt"],
        "properties": {
          "user": { "$ref": "#/components/schemas/AuthPrincipal" },
          "csrfToken": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43,
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "idleExpiresAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "absoluteExpiresAt": { "$ref": "#/components/schemas/UtcTimestamp" }
        }
      },
      "AuthError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "authentication_required",
                  "authentication_failed",
                  "permission_denied"
                ]
              }
            }
          }
        }
      },
...
      "CanonicalUuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      },
      "UtcTimestamp": {
        "type": "string",
        "format": "date-time",
        "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\\.[0-9]{6}Z$"
      },
...
      "PageCursor": {
        "type": "string",
        "minLength": 1,
        "maxLength": 2048,
        "pattern": "^[A-Za-z0-9_-]+$"
      },
      "InventoryResourceKind": {
        "type": "string",
        "enum": [
          "pve_cluster",
          "pve_node",
          "pve_guest",
          "pve_storage",
          "pbs_server",
          "pbs_datastore",
          "pbs_namespace",
          "pbs_backup_group",
          "pbs_snapshot"
        ]
      },
      "InventoryState": {
        "type": "string",
        "enum": ["active", "archived"],
        "default": "active"
      },
      "PageMetadata": {
        "type": "object",
        "additionalProperties": false,
        "required": ["limit", "count", "hasMore", "nextCursor"],
        "properties": {
          "limit": { "type": "integer", "minimum": 1, "maximum": 100 },
          "count": { "type": "integer", "minimum": 0, "maximum": 100 },
          "hasMore": { "type": "boolean" },
          "nextCursor": {
            "oneOf": [
              { "$ref": "#/components/schemas/PageCursor" },
              { "type": "null" }
            ]
          }
        }
      },
      "InventoryCounts": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "pveConnections",
          "pbsConnections",
          "pveClusters",
          "pveNodes",
          "qemuGuests",
          "lxcGuests",
          "pveStorages",
          "pbsServers",
          "pbsDatastores",
          "pbsNamespaces",
          "pbsBackupGroups",
          "pbsSnapshots"
        ],
        "properties": {
          "pveConnections": { "type": "integer", "minimum": 0 },
          "pbsConnections": { "type": "integer", "minimum": 0 },
          "pveClusters": { "type": "integer", "minimum": 0 },
          "pveNodes": { "type": "integer", "minimum": 0 },
          "qemuGuests": { "type": "integer", "minimum": 0 },
          "lxcGuests": { "type": "integer", "minimum": 0 },
          "pveStorages": { "type": "integer", "minimum": 0 },
          "pbsServers": { "type": "integer", "minimum": 0 },
          "pbsDatastores": { "type": "integer", "minimum": 0 },
          "pbsNamespaces": { "type": "integer", "minimum": 0 },
          "pbsBackupGroups": { "type": "integer", "minimum": 0 },
          "pbsSnapshots": { "type": "integer", "minimum": 0 }
        }
      },
      "InventoryOverview": {
        "type": "object",
        "additionalProperties": false,
        "required": ["generatedAt", "latestInventoryAt", "counts"],
        "properties": {
          "generatedAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "latestInventoryAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          },
          "counts": { "$ref": "#/components/schemas/InventoryCounts" }
        }
      },
...
      "InventoryResource": {
        "oneOf": [
          { "$ref": "#/components/schemas/PveClusterResource" },
          { "$ref": "#/components/schemas/PveNodeResource" },
          { "$ref": "#/components/schemas/PveGuestResource" },
          { "$ref": "#/components/schemas/PveStorageResource" },
          { "$ref": "#/components/schemas/PbsServerResource" },
          { "$ref": "#/components/schemas/PbsDatastoreResource" },
          { "$ref": "#/components/schemas/PbsNamespaceResource" },
          { "$ref": "#/components/schemas/PbsBackupGroupResource" },
          { "$ref": "#/components/schemas/PbsSnapshotResource" }
        ],
        "discriminator": { "propertyName": "kind" }
      },
      "InventoryResourcePage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page"],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": { "$ref": "#/components/schemas/InventoryResource" }
          },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        }
      },
...
      "CollectorScheduleUnconfigured": {
        "type": "object",
        "additionalProperties": false,
        "required": ["configured"],
        "properties": { "configured": { "const": false } }
      },
      "CollectorScheduleConfigured": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "configured",
          "intervalSeconds",
          "nextScanAt",
          "lastCycleStartedAt",
          "lastCycleFinishedAt",
          "leaseActive",
          "leaseExpiresAt"
        ],
        "properties": {
          "configured": { "const": true },
          "intervalSeconds": { "type": "integer", "minimum": 1 },
          "nextScanAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "lastCycleStartedAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          },
          "lastCycleFinishedAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          },
          "leaseActive": { "type": "boolean" },
          "leaseExpiresAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          }
        }
      },
      "CollectorHeartbeat": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "startedAt",
          "heartbeatAt",
          "expiresAt",
          "fresh",
          "currentActivity",
          "nextActionAt",
          "buildVersion"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["starting", "ready", "busy", "degraded", "stopping"]
          },
          "startedAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "heartbeatAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "expiresAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "fresh": { "type": "boolean" },
          "currentActivity": {
            "oneOf": [{ "type": "string" }, { "type": "null" }]
          },
          "nextActionAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          },
          "buildVersion": { "type": "string" }
        }
      },
      "CollectorStatus": {
        "type": "object",
        "additionalProperties": false,
        "required": ["generatedAt", "schedule", "heartbeat"],
        "properties": {
          "generatedAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "schedule": {
            "oneOf": [
              { "$ref": "#/components/schemas/CollectorScheduleUnconfigured" },
              { "$ref": "#/components/schemas/CollectorScheduleConfigured" }
            ]
          },
          "heartbeat": {
            "oneOf": [
              { "$ref": "#/components/schemas/CollectorHeartbeat" },
              { "type": "null" }
            ]
          }
        }
      },
      "CollectorRun": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "connectionId",
          "connectionName",
          "product",
          "status",
          "authoritative",
          "startedAt",
          "finishedAt",
          "appliedAt",
          "nodesSeen",
          "guestsSeen",
          "storagesSeen",
          "errorCode"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/CanonicalUuid" },
          "connectionId": { "$ref": "#/components/schemas/CanonicalUuid" },
          "connectionName": { "type": "string" },
          "product": { "type": "string", "enum": ["pve", "pbs"] },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "succeeded",
              "partial",
              "failed",
              "cancelled",
              "abandoned"
            ]
          },
          "authoritative": { "type": "boolean" },
          "startedAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "finishedAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          },
          "appliedAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          },
          "nodesSeen": { "type": "integer", "minimum": 0 },
          "guestsSeen": { "type": "integer", "minimum": 0 },
          "storagesSeen": { "type": "integer", "minimum": 0 },
          "errorCode": {
            "oneOf": [
              { "type": "string", "pattern": "^[a-z0-9_]{1,64}$" },
              { "type": "null" }
            ]
          }
        }
      },
      "CollectorScope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "runId",
          "scopeType",
          "scopeKey",
          "status",
          "observedAt",
          "errorCode"
        ],
        "properties": {
          "runId": { "$ref": "#/components/schemas/CanonicalUuid" },
          "scopeType": {
            "type": "string",
            "enum": [
              "pve_topology",
              "pve_guests",
              "pve_storages",
              "pve_node_storages",
              "pbs_system",
              "pbs_datastores",
              "pbs_datastore_status",
              "pbs_namespaces",
              "pbs_snapshots",
              "pve_backup_jobs",
              "pve_tasks_active",
              "pve_tasks_archive",
              "pbs_prune_jobs",
              "pbs_sync_jobs",
              "pbs_verify_jobs",
              "pbs_tasks_running",
              "pbs_tasks_window"
            ]
          },
          "scopeKey": { "type": "string", "maxLength": 512 },
          "status": {
            "type": "string",
            "enum": ["complete", "partial", "failed"]
          },
          "observedAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "errorCode": {
            "oneOf": [
              { "type": "string", "pattern": "^[a-z0-9_]{1,64}$" },
              { "type": "null" }
            ]
          }
        }
      },
      "CollectorRunPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page"],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": { "$ref": "#/components/schemas/CollectorRun" }
          },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        }
      },
      "CollectorScopePage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page"],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": { "$ref": "#/components/schemas/CollectorScope" }
          },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        }
      },
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message"],
            "properties": {
              "code": { "const": "invalid_query" },
              "message": { "const": "The query is invalid." }
            }
          }
        }
      },
...
    }
  }
}